refactor: move setup-token retirement into packages/api
Keep a thin wrapper in /api that injects getUserById. The lookup
and cutoff decision now live in typed backend code.
